requirements.txt
anyio==4.2.0
asgiref==3.7.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.1
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
pycryptodome==3.20.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
urllib3==2.1.0
Werkzeug==3.0.1
